added callback function for cache structures

+++ src/discord/disco.h
@@ -35,10 +35,10 @@ typedef struct disco_event_callbacks {
      * deleted, the message struct is also included.
      *
      */
-    void (*on_message_delete)(bot_client_t *, int64_t message_id, int64_t channel_id, int64_t guild_id, struct discord_message *);
+    void (*on_message_delete)(bot_client_t *, uint64_t message_id, uint64_t channel_id, uint64_t guild_id, struct discord_message *);
     void (*on_channel_create)(bot_client_t *, struct discord_channel *channel);
     void (*on_channel_update)(bot_client_t *, struct discord_channel *old, struct discord_channel *new);
-    void (*on_channel_delete)(bot_client_t *, int64_t channel_id, int64_t guild_id, int64_t parent_id, enum Discord_Channel_Type type, struct discord_channel *);
+    void (*on_channel_delete)(bot_client_t *, uint64_t channel_id, uint64_t guild_id, uint64_t parent_id, enum Discord_Channel_Type type, struct discord_channel *);
 } disco_event_callbacks_t;
 
 typedef struct bot_client {
@@ -87,6 +87,6 @@ int get_int_from_json(cJSON *data, const char *name, int default_);
 typedef void *(*disco_struct_fn)(cJSON *);
 int get_array_from_json(cJSON *data, const char *name, void ***array, size_t s, disco_struct_fn func);
 
-int64_t get_long_from_string_json(cJSON *data, const char *name, int default_);
+uint64_t get_long_from_string_json(cJSON *data, const char *name, uint64_t default_);
 
 #endif
